# openAPI_specification/OpenAPI.yaml
openapi: "3.0.3"
info:
  title: Dummy Roblox Shop
  version: "1.0"
  description: A Codecademy exercise building for fun very basic online e-store - Roblox Shop API
servers:
  - url: https://api.server.test/v1
...
security:
  - BasicAuth: []
paths:
...
  /products/product/:productId:
    summary: Get product data
    description: This path is used for retreiving a single product data
    get:
      tags:
        - Products
      summary: Get product details
      description: Retreive product data based on product ID
      operationId: get_product
      parameters:
        - $ref: "#/components/parameters/productIdPathParameter"
      responses:
        "200":
          description: Product data retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /cart:
    summary: Getting cart
    description: This path is used for retreiving cart details if user is logged in.
    get:
      tags:
        - Cart
      summary: Get user cart
      description: Retreive Cart data for logged in user.
      operationId: get_cart
      responses:
        "200":
          description: Cart data retreived successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  userid:
                    type: integer
                  Products:
                    type: array
                    items:
                      $ref: "#/components/schemas/CartProductWithQty"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /orders:
    summary: Retreive all User orders
    description: This path is used for retreiving all orders of the logged in user
    get:
      summary: Get all orders of User with order details
      description: Get all User Orders
      operationId: get_orders
      tags:
        - Orders
      responses:
        "200":
          description: All orders data retreived successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    userid:
                      type: integer
                    Products:
                      type: array
                      items:
                        $ref: "#/components/schemas/OrderProductWithQty"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /orders/order/:orderId:
    summary: Get specific orders
    description: This path is used for retreiving specific order based on passed ID
    get:
      summary: Get specific order details
      description: Get specific order based on order Id
      tags:
        - Orders
      operationId: get_order
      parameters:
        - $ref: "#/components/parameters/orderIdPathParameter"
      responses:
        "200":
          description: All orders data retreived successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  userid:
                    type: integer
                  Products:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrderProductWithQty"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /admin/products/addProduct:
    summary: Add produtct to DB
    description: This path is used for adding a new Product to the DB
    post:
      summary: Add product to DB
      description: Add passed object as product to the DB
      operationId: admin_add_product
      tags:
        - Admin
        - Products
      requestBody:
        description: Object with Porduct properties
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BaseProduct"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/BaseProduct"
      responses:
        "201":
          description: Product added to DB successfully, new product object retured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseProduct"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /admin/products/removeProduct:
    summary: Delete product from DB
    description: This path is used for deletion of product record from the DB
    delete:
      summary: Delete product from DB
      description: Delete product from DB based on the product ID passed
      operationId: admin_delete_product
      tags:
        - Admin
        - Products
      requestBody:
        description: request body should containt object with product Id
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductId"

      responses:
        "200":
          description: Product deleted successfully with confirmation message sent back as a responses object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationMessage"
  /admin/products/editProduct:
    summary: Edit product details
    description: This path is used for editing the product details
    put:
      summary: Update product details
      description: Updated product details based on the product object passed in the request body
      operationId: admin_update_product
      tags:
        - Admin
        - Products
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: Product details updated successfully. New product details object sent back.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /admin/orders/cancel:
    summary: Cancel existing order
    description: This path is used for canceling the existing orders
    delete:
      summary: Delete Order
      description: Deletes order based on order ID
      operationId: admin_order_delete
      tags:
        - Admin
        - Orders
      requestBody:
        description: Object with order ID
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderId"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/OrderId"
      responses:
        "204":
          description: Order deleted successfully with confirmation message sent back as a responses object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationMessage"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  parameters:
    productIdPathParameter:
      name: orderId
      description: Order ID number
      in: path
      required: true
      schema:
        type: integer
    orderIdPathParameter:
      name: orderId
      description: Order Id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ProductId:
      type: object
      properties:
        productId:
          type: integer
      required:
        - productId
    BaseProduct:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        category:
          type: string
          enum:
            [
              "general",
              "outerwear",
              "pants",
              "skirts",
              "headwear",
              "accessories",
              "shoes",
            ]
      required:
        - name
        - description
        - price
        - category
    Product:
      allOf:
        - type: object
          properties:
            id:
              type: integer
          required: [id]
        - $ref: "#/components/schemas/BaseProduct"
    CartProductWithQty:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            CartsProducts:
              type: object
              properties:
                qty:
                  type: integer
    OrderId:
      type: object
      properties:
        orderId:
          type: integer
      required:
        - orderId
    OrderProductWithQty:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            "OrdersProducts":
              type: object
              properties:
                qty:
                  type: integer
                  nullable: false
    OperationMessage:
      type: object
      properties:
        msg:
          description: Server response message
          type: string
...
  responses:
    InternalServerError:
      description: Error at server end.
      content:
        application/json:
          schema:
            type: object
            properties:
              msg:
                type: string
                description: Error message
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic